requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
fpdf2
